--- batch/README.md
@@ -1,9 +1,0 @@
-# batch
-
-Some util functions related to batching use cases.
-
-## Installation
-
-```shell
-go get github.com/that-happy-place/golean/batch@latest
-```
