--- batch/go.mod
@@ -1,3 +1,0 @@
-module github.com/that-happy-place/golean/batch
-
-go 1.21
